+++ packages/terra-core-docs/src/terra-dev-site/test/status-view/StatusViewAlignTop.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import StatusView from '../../../StatusView';
+import StatusView from 'terra-status-view';
 import styles from './StatusViewTestCommon.module.scss';
 
 const cx = classNames.bind(styles);